+++ src/bedrock/symbols/windows.h
@@ -11,7 +11,7 @@
 
 namespace endstone::runtime {
 
-static constexpr std::array<std::pair<std::string_view, std::size_t>, 61> symbols = {{
+static constexpr std::array<std::pair<std::string_view, std::size_t>, 63> symbols = {{
     {"BlockState::StateListNode::mHead", 211294960},
     {"Enchant::mEnchants", 211141912},
     // Actor
@@ -79,6 +79,8 @@ static constexpr std::array<std::pair<std::string_view, std::size_t>, 61> symbol
     {"?createPacket@MinecraftPackets@@SA?AV?$shared_ptr@VPacket@@@std@@W4MinecraftPacketIds@@@Z", 12540224},
     // Mob
     {"?knockback@Mob@@UEAAXPEAVActor@@MMMAEBUKnockbackParameters@@@Z", 37751424},
+    // NetherNet::SimpleNetworkInterfaceImpl
+    {"?ReceiveFromSignalingChannel@SimpleNetworkInterfaceImpl@NetherNet@@AEAAXUNetworkID@2@V?$basic_string_view@DU?$char_traits@D@std@@@std@@W4SignalingChannelId@2@@Z", 18932048},
     // NetworkSystem
     {"?send@NetworkSystem@@QEAAXAEBVNetworkIdentifier@@AEBVPacket@@W4SubClientId@@@Z", 12586720},
     // PistonBlockActor
@@ -115,6 +117,8 @@ static constexpr std::array<std::pair<std::string_view, std::size_t>, 61> symbol
     {"?updateWeather@WeatherManager@@QEAAXMHMH@Z", 19355200},
     // YAAEAVI18n
     {"?getI18n@@YAAEAVI18n@@XZ", 6057456},
+    // webrtc::P2PTransportChannel
+    {"?SwitchSelectedConnectionInternal@P2PTransportChannel@webrtc@@AEAAXPEAVConnection@2@W4IceSwitchReason@2@@Z", 170225584},
 }};
 
 }  // namespace endstone::runtime